refactor helper methods into kafka package for re-use in Ender
