+++ indexer/packages/kafka/src/index.ts
@@ -5,6 +5,7 @@ export * from './kafka';
 export * from './constants';
 export * from './types';
 export * from './batch-kafka-producer';
+export * from './websocket-helper';
 export { kafkaConfigSchema } from './config';
 
 export * from '../__tests__/helpers/kafka';
